cleanup and fix bugs

+++ src/tpConvolution.h
@@ -9,4 +9,4 @@ cv::Mat convolution(cv::Mat image, cv::Mat kernel);
 
 cv::Mat edgeSobel(cv::Mat image);
 
-cv::Mat bilateralFilter(cv::Mat image, cv::Mat kernel, double sigma_r);
+cv::Mat bilateralFilter(cv::Mat image, cv::Mat kernel, float sigma_r);
